Skip first inspection

+++ cv/actions/inspection_end.py
@@ -4,9 +4,13 @@ from config.plc_db import SIG_SEND_RESULT
 
 def inspection_end_actions(data):
     # Send data to PLC and save
-    data.artifact.end_inspection()
-    send_signal(SIG_SEND_RESULT, data.artifact.get_part_results_plc())
-    data.artifact.save()
+    if data.artifact is not None:
+        data.artifact.end_inspection()
+        send_signal(SIG_SEND_RESULT, data.artifact.get_part_results_plc())
+        data.artifact.save()
+
+    # Inspection is active after at least one cycle is completed
+    data.is_active = True
 
     # Reset Cycle
     data.artifact = None
